Docklike: Remove TaskbarFrame name in targets
When a monitor is attached and the `Taskbar.TaskbarFrame` element first
appears, it doesn't have a name yet, and so the target doesn't match.
Then, the name is added, but Windhawk doesn't track that, it only
evaluates the target once. Just use `Taskbar.TaskbarFrame` as a target
instead of `Taskbar.TaskbarFrame#TaskbarFrame`.