Requested changes - WIP1
May have errors, I'll review it again when I can.

- MultiWindowElement fix-ish
It's not looking very good, but I also I didn't had the same issue, not sure if it's because of the Windows version. I don't know how the fix is going to behave in other versions. I should be getting 25h2 soon; I had the option to update last week, but it was failing to install.

- I think all unstyled targets have been removed

- Removed Acrylic limitation from Known Issues